requirements.txt
streamlit
python-dotenv
openai
chromadb
langchain
langchain_chroma
langchain_openai
langchain_community
python3.11
matplotlib
pandas

